==> draft-project-2/Notebooks/03_wf_load_staging_area.ipynb <==
1 - IMPORT MAIN LIBRARIES AND CREATE GLOBAL VARIABLES

import pandas as pd
from sqlalchemy import create_engine

trusted_path = '..\\Trusted\\'

pg_connection_string = 'postgresql://admin:admin@localhost:5435/analytics'
pg_engine = create_engine(pg_connection_string)

2 - Load Trusted Bank Data

# Load csv trusted data to data frame
df_banks = pd.read_csv(trusted_path + 'banks.csv', encoding='cp1252')

3 - Load Trusted Taxes Data

# Load csv trusted data to data frame
df_taxes = pd.read_csv(trusted_path + 'taxes.csv', encoding='cp1252')

4 - CREATE STAGING AREA

# Run create schema SQL command
pg_engine.execute(
'''CREATE SCHEMA IF NOT EXISTS staging;'''
)

5 - CREATE BANK TABLE AND LOAD TRUSTED DATA

# Create table in postgres
pg_engine.execute(
'''DROP TABLE IF EXISTS staging.bancos;
CREATE TABLE IF NOT EXISTS staging.bancos (
    id serial PRIMARY KEY,
    created_at TIMESTAMP DEFAULT NOW(),
    ano INTEGER,
    trimestre VARCHAR (255),
    categoria VARCHAR (255),
    tipo VARCHAR (255),
    cnpj VARCHAR (255),
    instituicao_financeira VARCHAR (255),
    indice DECIMAL,
    qtde_reclamacoes_reguladas_procedentes INTEGER,
    qtde_reclamacoes_reguladas_outras INTEGER,
    qtde_de_reclamacoes_nao_reguladas INTEGER,
    qtde_total_reclamacoes INTEGER,
    qtde_total_clientes_spa_ccs_e_scr INTEGER,
    qtde_clientes_spa_ccs INTEGER,
    qtde_clientes_spa_scr INTEGER
);'''
)

# Load dataframe to postgres
df_banks.to_sql('staging.bancos', df_banks)

6 - CREATE TAXES TABLE AND LOAD TRUSTED DATA

# Create table in postgres
pg_engine.execute(
'''DROP TABLE IF EXISTS staging.lista_tarifas;
CREATE TABLE IF NOT EXISTS staging.lista_tarifas (
    id serial PRIMARY KEY,
    created_at TIMESTAMP DEFAULT NOW(),
    cnpj VARCHAR (255),
    codigo_servico INT,
    data_vigencia VARCHAR(255),
    periodicidade VARCHAR(255),
    servico VARCHAR(255),
    tipo_valor VARCHAR(255),
    unidade VARCHAR(255),
    valor_maximo DECIMAL(18,2)
);
''')

# Load dataframe to postgres
df_taxes.to_sql('staging.lista_tarifas', pg_engine)
