==> draft-project-2/Notebooks/05_data_visualization.ipynb <==
1 - IMPORT MAIN LIBRARIES AND CREATE GLOBAL VARIABLES

import pandas as pd
from sqlalchemy import create_engine
import seaborn as sns


# pg_connection_string = 'postgresql://admin:admin@localhost:5435/analytics'
# pg_engine = create_engine(pg_connection_string)
